refactor(Pawn): 폰 색별로 클래스 분리
